AWS/aula08/pyproject.toml
[tool.poetry]
name = "aula08"
version = "0.1.0"
description = ""
authors = ["Thiago Ferraz <thiagovdferraz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
schedule = "^1.2.2"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
